# backend/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recluse API",
    "description": "\nTwo-stage network intrusion detection.\n\n* **Stage 1** — supervised classifier: names the attack families it was trained on.\n* **Stage 2** — benign-only autoencoder: flags traffic that does not look like\n  normal, including families Stage 1 has never seen.\n\nThe system alerts, ranks and explains. It never blocks traffic.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/replay/start": {
      "post": {
        "tags": [
          "traffic"
        ],
        "summary": "Start replaying held-out test flows at 1x, 10x or 100x",
        "description": "Begin streaming a held-out split, scoring in batches and pushing over SSE.\n\n**202, not 200.** The work this starts outlives the request: the response\nsays the replay has been accepted and is now running, not that it has\nfinished. A 200 would claim a completed job.",
        "operationId": "replay_start_api_v1_replay_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayStatus"
                }
              }
            }
          },
          "409": {
            "description": "A replay is already running."
          },
          "422": {
            "description": "Unknown dataset, or a speed outside 1 / 10 / 100."
          },
          "503": {
            "description": "No model bundle is loaded, so there is nothing to score with."
          }
        }
      }
    },
    "/api/v1/replay/stop": {
      "post": {
        "tags": [
          "traffic"
        ],
        "summary": "Stop the active replay",
        "description": "Cancel the running replay and report what it managed before stopping.\n\nThe counters in the response are the run's final totals, which is the\nuseful thing to return from a stop -- \"it scored 142,000 rows and raised\n318 alerts before you stopped it\" rather than an empty acknowledgement.",
        "operationId": "replay_stop_api_v1_replay_stop_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayStatus"
                }
              }
            }
          },
          "409": {
            "description": "No replay is running."
          }
        }
      }
    },
...
    "/api/v1/ingest/start": {
      "post": {
        "tags": [
          "traffic"
        ],
        "summary": "Begin scoring a live capture (authorised networks only)",
        "description": "Start a shadow burn-in or a live alerting capture.\n\n**Where it runs is configuration, not a parameter.** The interface must be\none the operator listed in `IDS_LIVE_INTERFACES` -- capture is lawful only on\nnetworks you own or are authorised to monitor -- and a pcap must be a file in\n`IDS_LIVE_PCAP_DIR`. **Shadow before alert:** alert mode needs a local\n`tau_anom` from `python -m training.calibrate_live`, because the CICIDS2017\nthreshold on a real network floods the queue with its ordinary traffic.",
        "operationId": "ingest_start_api_v1_ingest_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStatus"
                }
              }
            }
          },
          "403": {
            "description": "The interface is not in IDS_LIVE_INTERFACES."
          },
          "404": {
            "description": "No such pcap in IDS_LIVE_PCAP_DIR."
          },
          "409": {
            "description": "A capture or replay is running, or alert mode was asked for before a shadow burn-in calibrated a local threshold."
          },
          "422": {
            "description": "An invalid source, mode, or a missing interface or pcap name."
          },
          "503": {
            "description": "Stage 2 is not loaded."
          }
        }
      }
    },
    "/api/v1/ingest/stop": {
      "post": {
        "tags": [
          "traffic"
        ],
        "summary": "Stop the live capture after scoring what it has already metered",
        "description": "Stop capturing; flows still open are finished and scored before this returns.",
        "operationId": "ingest_stop_api_v1_ingest_stop_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStatus"
                }
              }
            }
          },
          "409": {
            "description": "No capture is running."
          }
        }
      }
    },
    "/api/v1/stream": {
      "get": {
        "tags": [
          "stream"
        ],
        "summary": "Live alert feed over server-sent events",
        "description": "Hold a connection open and push one event per persisted alert.\n\nAnswers **503** when no traffic source is running rather than holding open\na connection that can never produce an event. An empty stream and a stream\nwith nothing to say are indistinguishable from the client's side, and the\nfirst is a bug while the second is Tuesday -- so the two are given\ndifferent status codes. `POST /replay/start` is what makes a source\nactive.",
        "operationId": "stream_alerts_api_v1_stream_get",
        "responses": {
          "200": {
            "description": "An open `text/event-stream` of `alert` and `heartbeat` events.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AlertEvent"
                    },
                    {
                      "$ref": "#/components/schemas/HeartbeatEvent"
                    }
                  ],
                  "type": "string",
                  "title": "Response 200 Stream Alerts Api V1 Stream Get"
                }
              }
            }
          },
          "503": {
            "description": "No traffic source is running, so there is nothing to stream."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AlertEvent": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "KNOWN",
              "UNCLASSIFIED_ANOMALY"
            ],
            "title": "Kind"
          },
          "family": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "dos",
                  "ddos",
                  "brute_force",
                  "port_scan",
                  "web_attack",
                  "botnet",
                  "infiltration"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "anomaly_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomaly Score"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          },
          "src_ip": {
            "type": "string",
            "title": "Src Ip"
          },
          "dst_ip": {
            "type": "string",
            "title": "Dst Ip"
          },
          "occurrence_count": {
            "type": "integer",
            "title": "Occurrence Count"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "family",
          "severity",
          "risk_score",
          "anomaly_score",
          "detected_at",
          "src_ip",
          "dst_ip",
          "occurrence_count",
          "model_version"
        ],
        "title": "AlertEvent",
        "description": "Payload of the ``alert`` SSE event (API-Reference.md \"Planned event shape\").\n\nSSE responses are ``text/event-stream``, so FastAPI never attaches this\nas a ``response_model`` -- it exists purely so the frontend's\n``EventSource`` handler has a generated type instead of a hand-written\none that can silently drift from it. Deliberately narrower than\nAlertSummary: the ticker renders one line and the drawer fetches the\nrest when clicked."
      },
...
      "HeartbeatEvent": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "ts"
        ],
        "title": "HeartbeatEvent",
        "description": "Payload of the ``heartbeat`` SSE event.\n\nSent so a client that has not seen an alert in a while can still tell the\nconnection is alive rather than stalled."
      },
      "IngestStartRequest": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "interface",
              "pcap"
            ],
            "title": "Source"
          },
          "interface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface",
            "description": "An interface listed in IDS_LIVE_INTERFACES."
          },
          "pcap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pcap",
            "description": "A file name in IDS_LIVE_PCAP_DIR."
          },
          "mode": {
            "type": "string",
            "enum": [
              "shadow",
              "alert"
            ],
            "title": "Mode",
            "description": "shadow scores and alerts no one (the burn-in); alert needs a local tau_anom calibrated for the serving Stage 2 model.",
            "default": "shadow"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "IngestStartRequest",
        "description": "POST /api/v1/ingest/start's request body (Phase 9).\n\nWhere a capture may run is not chosen here: `interface` must be one the\noperator listed in IDS_LIVE_INTERFACES, and `pcap` a file in\nIDS_LIVE_PCAP_DIR. Anything else is refused with 403 or 404."
      },
      "IngestStatus": {
        "properties": {
          "running": {
            "type": "boolean",
            "title": "Running"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "interface:<name> or pcap:<file>; null if never run."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "packets": {
            "type": "integer",
            "title": "Packets"
          },
          "undecoded": {
            "type": "integer",
            "title": "Undecoded",
            "description": "Frames that were not TCP or UDP over IP."
          },
          "flows": {
            "type": "integer",
            "title": "Flows",
            "description": "Flows the meter finished."
          },
          "unscoreable": {
            "type": "integer",
            "title": "Unscoreable",
            "description": "Zero-duration and single-packet flows: infinite rates, never seen in training."
          },
          "scored": {
            "type": "integer",
            "title": "Scored"
          },
          "shadow_rows": {
            "type": "integer",
            "title": "Shadow Rows"
          },
          "alerts": {
            "type": "integer",
            "title": "Alerts",
            "description": "Flows that raised an alert, before dedupe folds them."
          },
          "tau_anom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tau Anom",
            "description": "The Stage 2 threshold this capture decides at."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "allowed_interfaces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Interfaces",
            "description": "IDS_LIVE_INTERFACES, as configured."
          },
          "pcaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pcaps",
            "description": "Files in IDS_LIVE_PCAP_DIR."
          },
          "tau_anom_dataset": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tau Anom Dataset"
          },
          "calibration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalCalibration"
              },
              {
                "type": "null"
              }
            ],
            "description": "The usable local calibration, or null if none matches the serving model."
          }
        },
        "type": "object",
        "required": [
          "running",
          "mode",
          "source",
          "session_id",
          "started_at",
          "finished_at",
          "packets",
          "undecoded",
          "flows",
          "unscoreable",
          "scored",
          "shadow_rows",
          "alerts",
          "tau_anom",
          "error",
          "allowed_interfaces",
          "pcaps",
          "tau_anom_dataset",
          "calibration"
        ],
        "title": "IngestStatus",
        "description": "GET /api/v1/ingest/status, and what /ingest/start and /ingest/stop return."
      },
      "LocalCalibration": {
        "properties": {
          "tau_anom_local": {
            "type": "number",
            "title": "Tau Anom Local"
          },
          "tau_anom_dataset": {
            "type": "number",
            "title": "Tau Anom Dataset"
          },
          "percentile": {
            "type": "number",
            "title": "Percentile"
          },
          "flows": {
            "type": "integer",
            "title": "Flows",
            "description": "Shadow-scored flows the percentile was cut from."
          },
          "computed_at": {
            "type": "string",
            "title": "Computed At"
          },
          "stage2_version": {
            "type": "string",
            "title": "Stage2 Version"
          },
          "dataset_threshold_alert_rate": {
            "type": "number",
            "title": "Dataset Threshold Alert Rate",
            "description": "Share of the burn-in the CICIDS2017 tau_anom would have flagged."
          },
          "stage1_alert_rate": {
            "type": "number",
            "title": "Stage1 Alert Rate",
            "description": "Share of the burn-in Stage 1 would have named at tau_sup."
          },
          "capture_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capture Sources"
          },
          "window_start": {
            "type": "string",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "title": "Window End"
          }
        },
        "type": "object",
        "required": [
          "tau_anom_local",
          "tau_anom_dataset",
          "percentile",
          "flows",
          "computed_at",
          "stage2_version",
          "dataset_threshold_alert_rate",
          "stage1_alert_rate",
          "capture_sources",
          "window_start",
          "window_end"
        ],
        "title": "LocalCalibration",
        "description": "The local tau_anom a shadow burn-in produced (training.calibrate_live)."
      },
...
      "ReplayStartRequest": {
        "properties": {
          "speed": {
            "type": "integer",
            "enum": [
              1,
              10,
              100
            ],
            "title": "Speed",
            "description": "Time acceleration factor."
          },
          "dataset": {
            "type": "string",
            "title": "Dataset",
            "description": "Held-out split name to replay."
          }
        },
        "type": "object",
        "required": [
          "speed",
          "dataset"
        ],
        "title": "ReplayStartRequest",
        "description": "POST /api/v1/replay/start's request body.\n\n``speed`` is a ``Literal`` of exactly the documented allowed values, so an\nunlisted speed is a 422 from the validation layer rather than something\nthe replay engine has to police."
      },
      "ReplayStatus": {
        "properties": {
          "running": {
            "type": "boolean",
            "title": "Running"
          },
          "speed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed",
            "description": "Null when no replay is running."
          },
          "dataset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset",
            "description": "Null when no replay is running."
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "Null when no replay is running."
          },
          "rows_scored": {
            "type": "integer",
            "title": "Rows Scored"
          },
          "alerts_emitted": {
            "type": "integer",
            "title": "Alerts Emitted"
          }
        },
        "type": "object",
        "required": [
          "running",
          "speed",
          "dataset",
          "started_at",
          "rows_scored",
          "alerts_emitted"
        ],
        "title": "ReplayStatus",
        "description": "What POST /api/v1/replay/start and /replay/stop return on 202."
      },
...
    }
  }
}